Source/Services/Tailwind.Traders.Cognitive.Docker/app/requirements.txt
tensorflow==1.5.0
pillow==5.0.0
numpy==1.14.1
flask==0.12.2